add authentication

+++ src/controllers/productsController.js
@@ -1,14 +1,14 @@
 import createHttpError from "http-errors";
 import { Product } from "../models/product.js"
 
-export const getProducts = async(req, res) => {
-    const products = await Product.find();
+export const getProducts = async (req, res) => {
+    const products = await Product.find({userId: req.user._id});
     res.status(200).json(products);
 }
 
 export const getProductById = async(req, res) => {
     const { productId } = req.params;
-    const product = await Product.findOne({ _id: productId });
+    const product = await Product.findOne({ _id: productId, userId: req.user._id });
 
     if (!product) {
         throw createHttpError(404, "Product not found");
@@ -17,13 +17,13 @@ export const getProductById = async(req, res) => {
 }
 
 export const createProduct = async(req, res) => {
-    const product = await Product.create(req.body);
+    const product = await Product.create({ ...req.body, userId: req.user._id });
     res.status(201).json(product);
 }
 
 export const updateProduct = async(req, res) => {
     const { productId } = req.params;
-    const product = await Product.findOneAndUpdate({ _id: productId }, req.body, { returnDocument: "after" });
+    const product = await Product.findOneAndUpdate({ _id: productId, userId: req.user._id }, req.body, { returnDocument: "after" });
     if (!product) {
         throw createHttpError(404, "Product not found");
     }
@@ -32,7 +32,7 @@ export const updateProduct = async(req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const { productId } = req.params;
-    const product = await Product.findOneAndDelete({ _id: productId });
+    const product = await Product.findOneAndDelete({ _id: productId, userId: req.user._id });
     if (!product) {
         throw createHttpError(404, "Product not found");
     }
